Changed to port F, program is set to go to alarm automatically

+++ Lab3Starter-master/src/AlarmClockSwitchControl.h
@@ -58,5 +58,5 @@ void switchDisplay(void);
 
 void runClockFace(void);
 void runDigitalClock(void);
-void settingTime(int32_t hour, int32_t min, uint8_t forAlarm);
+void settingTime(uint16_t hour, uint16_t min, uint8_t forAlarm);
 void checkAlarm(void);
